pyproject.toml
[tool.poetry]
name = "pibot"
version = "0.1.0"
description = ""
authors = ["Tom Haynes"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.2, <3.10"
tensorflow = "^2.11.0"
opencv-python = "^4.8.0.76"
scikit-learn = "^1.3.0"
numpy = "^1.26.0"
tensorflow-io-gcs-filesystem = "0.31.0"
pillow = "^10.0.1"

[tool.poetry.group.rpi]
optional = true

[tool.poetry.group.rpi.dependencies]
"RPi.GPIO" = "^0.7.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pytest = "^7.4.2"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
exclude = "RaspberryPi-Car, opencv-4.5.3"

[tool.isort]
skip = ["RaspberryPi-Car", "opencv-4.5.3"]
